// src/client/schemas.ts
export enum ClientCloseCode {
  CLIENT_CLOSE_CODE_UNSPECIFIED = 0,
  CLIENT_CLOSE_CODE_NORMAL = 1000,
  CLIENT_CLOSE_CODE_INTERNAL_SERVER_ERROR = 1011,
  CLIENT_CLOSE_CODE_STREAM_END = 4005,
  CLIENT_CLOSE_CODE_NO_MESSAGES_TIMEOUT = 4006,
  CLIENT_CLOSE_CODE_INVALID_OPTIONS = 4400,
  CLIENT_CLOSE_CODE_INVALID_AUTH = 4401,
  CLIENT_CLOSE_CODE_NO_PERMISSION = 4403,
  CLIENT_CLOSE_CODE_NOT_LIVE = 4404,
  CLIENT_CLOSE_CODE_TOO_MANY_CONNECTIONS = 4429,
  CLIENT_CLOSE_CODE_TIKTOK_CLOSED_CONNECTION = 4500,
  CLIENT_CLOSE_CODE_MAX_LIFETIME_EXCEEDED = 4555,
  CLIENT_CLOSE_CODE_WEBCAST_FETCH_ERROR = 4556,
  CLIENT_CLOSE_CODE_ROOM_INFO_FETCH_ERROR = 4557,
  UNRECOGNIZED = -1,
}